third_party/WebKit/LayoutTests/fast/scrolling/scroll-into-view-hidden-element.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div style="height: 2000px"></div>
<div style="overflow: scroll; height: 100px">
  <div id="scroll" style="margin-top: -15px; height: 15px; width: 15px;background-color: red"></div>
  Target
</div>
<script>
test(function(t) {
  var scroll = document.getElementById("scroll");
  scroll.scrollIntoView();
  assert_not_equals(window.scrollY, 0);
}, "Tests that when the element to be scrolled can't be scrolled to visible in its closest scrollable parent, the element's position should still be scrolled into view");
</script>
